Added `ipaddress` dep and condition on failure. Proxy uses `ipaddress` to check if the host is an IP. You can control how Innkeeper will fail when `check_in` does not return anything.
`config.redirect_on_failure` will accept one of the following:
* `:host_domain` -- Will redirect to the host domain (e.g. test.example.com fails to example.com)
* `:host_domain_with_www` -- Will redirect to the host domain with www in front of it (e.g. test.example.com fails to www.example.com)
* `:failure_app` -- With call the failure app defined in `config.failure_app`
* An instance of `String` containing the URL to which you want to redirect

If `config.failure_app` is not defined then Innkeeper will fail silently
by calling the next app in the middleware stack.

If using a `String` instance in `config.redirect_on_failure`, the `String`
is passed to rack unmodified, so make sure that it is properly formatted.